Split command_test.go by command (plan/apply/dump)
command_test.go had grown to 897 lines covering three separate CLI commands. Split it into plan_test.go, apply_test.go, and dump_test.go to match the existing per-command source layout (plan.go/apply.go/dump.go), with shared assertions extracted to helper_test.go. WriteDumpFiles tests stay with dump_test.go since the helper is dump-only.
