+++ cmd/command/export_test.go
@@ -1,4 +1,4 @@
 package command
 
-// Exported to tests in cmd/command/command_test.go (package command_test).
+// Exported to tests in cmd/command/dump_test.go (package command_test).
 var WriteDumpFiles = writeDumpFiles
